fix the whole capturer thing (code review)
