+++ internal/apmcontext/context.go
@@ -58,8 +58,6 @@ var (
 
 type spanKey struct{}
 type transactionKey struct{}
-type bodyCapturerKey struct{}
-type requestKey struct{}
 
 // DefaultContextWithSpan is the default value for ContextWithSpan.
 func DefaultContextWithSpan(ctx context.Context, span interface{}) context.Context {
@@ -79,17 +77,4 @@ func DefaultSpanFromContext(ctx context.Context) interface{} {
 // DefaultTransactionFromContext is the default value for TransactionFromContext.
 func DefaultTransactionFromContext(ctx context.Context) interface{} {
 	return ctx.Value(transactionKey{})
-}
-
-// ContextWithBodyCapturer takes a context and request body capturer and returns
-// a new context from which the capturer can be extracted using
-// BodyCapturerFromContext.
-func ContextWithBodyCapturer(ctx context.Context, bc interface{}) context.Context {
-	return context.WithValue(ctx, bodyCapturerKey{}, bc)
-}
-
-// BodyCapturerFromContext returns a request body capturer included in the context
-// using ContextWithBodyCapturer.
-func BodyCapturerFromContext(ctx context.Context) interface{} {
-	return ctx.Value(bodyCapturerKey{})
 }
