+++ module/apmechov4/middleware.go
@@ -70,11 +70,9 @@ func (m *middleware) handle(c echo.Context) error {
 		return m.handler(c)
 	}
 	name := req.Method + " " + c.Path()
-	tx, req := apmhttp.StartTransaction(m.tracer, name, req)
+	tx, body, req := apmhttp.StartTransactionWithBody(m.tracer, name, req)
 	defer tx.End()
 	c.SetRequest(req)
-	body := m.tracer.CaptureHTTPRequestBody(req)
-	req.WithContext(apm.ContextWithBodyCapturer(req.Context(), body))
 
 	resp := c.Response()
 	var handlerErr error
